modifyCarAdvert changes

+++ src/main/java/com/cars/adverts/services/CarService.java
@@ -55,10 +55,13 @@ public class CarService {
     //modify an existing car advert
     public Optional<CarAdvert> modifyCarAdvert(CarAdvertDTO carAdvertDTO, Long carId) {
         CarAdvert carAdvert = carAdvertRepository.findById(carId).get();
-        if(carAdvertDTO.getId()!=null)
+        if(carAdvertDTO.getId()!=null && carAdvertDTO.getId()>0)
         {
             carAdvertRepository.deleteById(carId);
 
+        }
+        else if(carAdvertDTO.getId()<0 ){
+            throw new RuntimeException("Id must be a positive number") ;
         }
         carAdvertMapper.updateCarAdvertFromDto(carAdvertDTO, carAdvert);
         carAdvertRepository.save(carAdvert);
